requirements.txt
chainlit
torch
transformers
python-dotenv
-e .